requirements.txt
# Core ML libraries
torch>=1.7.0
torchvision>=0.8.1
timm>=0.6.0

# YOLOv5 dependencies
ultralytics
opencv-python>=4.1.1
Pillow>=7.1.2
PyYAML>=5.3.1
requests>=2.23.0
scipy>=1.4.1
tqdm>=4.64.0

# Data science
matplotlib>=3.2.2
numpy>=1.18.5
seaborn>=0.11.0
pandas>=1.1.4

# Optional: Experiment tracking
# wandb
# tensorboard
